Rename subproject and package to refined
