--- refine/src/main/scala/monocle/refine/All.scala
@@ -1,5 +1,0 @@
-package monocle.refine
-
-object all extends RefineInstances
-
-trait RefineInstances extends BitsInstances
