+++ refined/src/main/scala/monocle/refined/All.scala
@@ -1,0 +1,5 @@
+package monocle.refined
+
+object all extends RefinedInstances
+
+trait RefinedInstances extends BitsInstances
