+++ refined/src/main/scala/monocle/refined/package.scala
@@ -4,7 +4,7 @@ import eu.timepit.refined._
 import eu.timepit.refined.api.Refined
 import eu.timepit.refined.numeric.Interval
 
-package object refine {
+package object refined {
   type ZeroTo[T] = Int Refined Interval[W.`0`.T, T]
 
   type ByteBits = ZeroTo[W.`7`.T]
