+++ test/src/test/scala/monocle/MonocleSuite.scala
@@ -2,7 +2,7 @@ package monocle
 
 import monocle.function.GenericOptics
 import monocle.generic.GenericInstances
-import monocle.refine.RefineInstances
+import monocle.refined.RefinedInstances
 import monocle.state.StateLensSyntax
 import monocle.std.StdInstances
 import monocle.syntax.Syntaxes
@@ -16,6 +16,6 @@ trait MonocleSuite extends FunSuite
                       with StdInstances
                       with GenericOptics
                       with GenericInstances
-                      with RefineInstances
+                      with RefinedInstances
                       with Syntaxes
                       with StateLensSyntax
